requirements.txt
fastapi==0.121.0
uvicorn[standard]==0.32.0
pandas==2.2.3
scikit-learn==1.6.1
mlflow==3.11.1
kagglehub==1.0.0
openai==1.58.1
cloudpickle==3.1.0
PyYAML==6.0.2
pytest==8.3.4
httpx==0.28.1
